Refactor prep_fastq_folder_ln to support dry-run dictionary mode

+++ rules/singlecell_import.smk
@@ -209,14 +209,12 @@ def filterFastq4nopipe(wildcards):
 
 record_fastqpath = {}
 record_fastqfiles = defaultdict(set)
-def prep_fastq_folder_ln(sample):
+def prep_fastq_folder_ln(sample, get_dict_only=False):
     """
     Prepare the folders for fastq symlinks and record the paths and files used.
-    Updates the global record_fastqpath and record_fastqfiles dictionaries.
+    Updates the global record_fastqpath and record_fastqfiles dictionaries only if get_dict_only is False.
     """
-    print("Preparing FASTQ folder for sample:", sample)
     global record_fastqpath, record_fastqfiles
-    # Check for both possible sample folder structures
     sample_folder_with_prefix = f"Sample_{sample}"
     sample_folder_without_prefix = sample
 
@@ -233,22 +231,20 @@ def prep_fastq_folder_ln(sample):
         sys.stderr.write(f"\nError: No FASTQ folder found for sample {detected_sample_folder}. Check the directory structure.\n\n")
         sys.exit(1)
 
-    # Define new FASTQ output directory
-    path_fq_new = os.path.join(analysis, 
-                        f"fastq/{detected_sample_folder}/")
-
-    # Remove existing files in the directory before creating symlinks
-    if os.path.exists(path_fq_new):
-        for file in os.listdir(path_fq_new):
-            file_path = os.path.join(path_fq_new, file)
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-    else:
-        os.makedirs(path_fq_new, exist_ok=True)
+    path_fq_new = os.path.join(analysis, f"fastq/{detected_sample_folder}/")
+    if not get_dict_only:
+        if os.path.exists(path_fq_new):
+            for file in os.listdir(path_fq_new):
+                file_path = os.path.join(path_fq_new, file)
+                if os.path.islink(file_path) or os.path.isfile(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+        else:
+            os.makedirs(path_fq_new, exist_ok=True)
 
     cnt_fq_file = 0
+    temp_fastqfiles = set()
     for index, fq_path in enumerate(fastqpath):
         path_sample = os.path.join(fq_path, detected_sample_folder)
         for fastq_file in glob.glob(os.path.join(path_sample, "*fastq.gz")):
@@ -264,24 +260,23 @@ def prep_fastq_folder_ln(sample):
             basename_fastq_new = f"{run_names_orig[index]}_{basename_fastq}"
             symlink_path = os.path.join(path_fq_new, basename_fastq_new)
             cmd = f"ln -s {fastq_file} {symlink_path}"
-            print(cmd)
-            process = Popen(cmd, shell=True, stdout=PIPE, stderr=PIPE)
-            process.wait()
-
-            # Record the symlinked file for this sample
-            record_fastqfiles[sample].add(symlink_path)
+            if not get_dict_only:
+                process = Popen(cmd, shell=True, stdout=PIPE, stderr=PIPE)
+                process.wait()
+            temp_fastqfiles.add(symlink_path)
 
     if cnt_fq_file == 0:
         sys.stderr.write("\nNo FASTQ files detected. Please check it out!\n\n")
         sys.exit(1)
 
-    # Record the output path for this sample
-    record_fastqpath[sample] = path_fq_new
-    #print(record_fastqpath)
+    if get_dict_only:
+        record_fastqpath[sample] = path_fq_new
+        record_fastqfiles[sample] = temp_fastqfiles
+
     return path_fq_new
 
 for sample in samples:
-    prep_fastq_folder_ln(sample)  # Prepare the fastq folder for each sample
+    prep_fastq_folder_ln(sample, get_dict_only=True)  # Prepare the fastq folder for each sample
 
 #Setting aggregate flag, this gets turned off for certain pipelines
 aggregate = True
